+++ test/integration/test_swap_with_claim_path.js
@@ -103,6 +103,10 @@ test(`Swap with claim path`, async ({end, equal, strictSame}) => {
                 return cbk({[args.name]: args.default});
               }
 
+              if (args.name === 'incoming') {
+                return cbk({incoming: false});
+              }
+
               if (args.name === 'req') {
                 return cbk({req: swapRequest.swap_request});
               }