+++ test/integration/test_swap_with_refund_path.js
@@ -104,6 +104,10 @@ test(`Timeout a swap`, async ({end, equal, strictSame}) => {
                   return cbk({[args.name]: args.default});
                 }
 
+                if (args.name === 'incoming') {
+                  return cbk({incoming: false});
+                }
+
                 if (args.name === 'req') {
                   return cbk({req: swapRequest.swap_request});
                 }